Renamed widgets subpackage to visualization and restructured modules within it (#1462)
* Reorganized widgets & tools modules

* Changed import paths as per new subpackage structure

* Updated imports in demo notebooks

* Added EOF newline to init files
