+++ tardis/visualization/__init__.py
@@ -1,0 +1,8 @@
+"""Visualization tools and widgets for TARDIS."""
+
+from tardis.visualization.widgets.shell_info import (
+    shell_info_from_simulation,
+    shell_info_from_hdf,
+)
+from tardis.visualization.widgets.line_info import LineInfoWidget
+from tardis.visualization.tools.sdec_plot import SDECPlotter
