+++ tardis/visualization/tools/__init__.py
@@ -1,0 +1,6 @@
+"""
+Classes & functions to generate diagnostic visualizations.
+
+These interfaces (or tools) post-process TARDIS simulations allowing user to
+produce a specific visualization through a function (or class method) call.
+"""
