--- tardis/widgets/__init__.py
@@ -1,8 +1,0 @@
-"""Visualization tools and widgets for TARDIS."""
-
-from tardis.widgets.shell_info import (
-    shell_info_from_simulation,
-    shell_info_from_hdf,
-)
-from tardis.widgets.line_info import LineInfoWidget
-from tardis.widgets.sdec_plot import SDECPlotter
